Add java docs for types
