+++ src/main/java/io/github/mapepire/types/ExplainType.java
@@ -1,6 +1,16 @@
 package io.github.mapepire.types;
 
+/**
+ * Enum representing the possible types of explain requests.
+ */
 public enum ExplainType {
+    /**
+     * Run the SQL statement for explanation.
+     */
     Run,
+
+    /**
+     * Do not run the SQL statement for explanation.
+     */
     DoNotRun
 }
