+++ src/main/java/io/github/mapepire/types/TransactionEndType.java
@@ -1,6 +1,16 @@
 package io.github.mapepire.types;
 
+/**
+ * Enum representing the possible types of transaction endings.
+ */
 public enum TransactionEndType {
+    /**
+     * Commit the transaction.
+     */
     COMMIT,
+
+    /**
+     * Rollback the transaction.
+     */
     ROLLBACK
 }
