+++ src/main/java/io/github/mapepire/types/jdbcOptions/Access.java
@@ -21,6 +21,7 @@ public enum Access {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
