+++ src/main/java/io/github/mapepire/types/jdbcOptions/BlockSize.java
@@ -26,6 +26,7 @@ public enum BlockSize {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
