+++ src/main/java/io/github/mapepire/types/jdbcOptions/CursorSensitivity.java
@@ -21,6 +21,7 @@ public enum CursorSensitivity {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
