+++ src/main/java/io/github/mapepire/types/jdbcOptions/DateSeparator.java
@@ -23,6 +23,7 @@ public enum DateSeparator {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
