+++ src/main/java/io/github/mapepire/types/jdbcOptions/Driver.java
@@ -20,6 +20,7 @@ public enum Driver {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
