+++ src/main/java/io/github/mapepire/types/jdbcOptions/Error.java
@@ -20,6 +20,7 @@ public enum Error {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
