+++ src/main/java/io/github/mapepire/types/jdbcOptions/MaximumPrecision.java
@@ -20,6 +20,7 @@ public enum MaximumPrecision {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
