+++ src/main/java/io/github/mapepire/types/jdbcOptions/MinimumDivideScale.java
@@ -28,6 +28,7 @@ public enum MinimumDivideScale {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
