+++ src/main/java/io/github/mapepire/types/jdbcOptions/Naming.java
@@ -20,6 +20,7 @@ public enum Naming {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
