+++ src/main/java/io/github/mapepire/types/jdbcOptions/PackageError.java
@@ -21,6 +21,7 @@ public enum PackageError {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
