+++ src/main/java/io/github/mapepire/types/jdbcOptions/QueryTimeoutMechanism.java
@@ -20,6 +20,7 @@ public enum QueryTimeoutMechanism {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
