+++ src/main/java/io/github/mapepire/types/jdbcOptions/Remarks.java
@@ -20,6 +20,7 @@ public enum Remarks {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
