+++ src/main/java/io/github/mapepire/types/jdbcOptions/ServerTrace.java
@@ -25,6 +25,7 @@ public enum ServerTrace {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
