+++ src/main/java/io/github/mapepire/types/jdbcOptions/Sort.java
@@ -21,6 +21,7 @@ public enum Sort {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
