+++ src/main/java/io/github/mapepire/types/jdbcOptions/SortWeight.java
@@ -20,6 +20,7 @@ public enum SortWeight {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
