+++ src/main/java/io/github/mapepire/types/jdbcOptions/TimeFormat.java
@@ -23,6 +23,7 @@ public enum TimeFormat {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
