+++ src/main/java/io/github/mapepire/types/jdbcOptions/TimeSeparator.java
@@ -22,6 +22,7 @@ public enum TimeSeparator {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
