+++ src/main/java/io/github/mapepire/types/jdbcOptions/ToolboxTrace.java
@@ -31,6 +31,7 @@ public enum ToolboxTrace {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
