+++ src/main/java/io/github/mapepire/types/jdbcOptions/TransactionIsolation.java
@@ -23,6 +23,7 @@ public enum TransactionIsolation {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
