+++ src/main/java/io/github/mapepire/types/jdbcOptions/TranslateHex.java
@@ -20,6 +20,7 @@ public enum TranslateHex {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
