+++ src/main/java/io/github/mapepire/types/jdbcOptions/XALooselyCoupledSupport.java
@@ -20,6 +20,7 @@ public enum XALooselyCoupledSupport {
                 return type;
             }
         }
+
         throw new IllegalArgumentException("Unknown value: " + value);
     }
 }
