+++ src/test/java/io/github/mapepire/PoolTest.java
@@ -10,6 +10,7 @@ import java.util.concurrent.CompletableFuture;
 
 import org.junit.jupiter.api.Test;
 
+import io.github.mapepire.types.PoolOptions;
 import io.github.mapepire.types.QueryResult;
 
 class PoolTest extends MapepireTest {
